+++ apps/portals/adknowledgeportal/src/config/style/_variable_overrides.scss
@@ -1,5 +1,5 @@
 @use '@sage-bionetworks/synapse-portal-framework/style/variables' with (
-  $primary-action-color: #4d5491,
+  $primary-action-color: #5a488f,
   $secondary-action-color: #2f8e94,
   $header-url: 'config/style/header.svg',
   $footer-url: 'config/style/header.svg'
